Use short for reference count in ContinuableScope

+++ dd-trace-core/src/main/java/datadog/trace/core/scopemanager/ContinuableScopeManager.java
@@ -181,7 +181,7 @@ public class ContinuableScopeManager implements AgentScopeManager {
 
     private final byte source;
 
-    private int referenceCount = 1;
+    private short referenceCount = 1;
 
     private final DDScopeEvent event;
